64 bit compilation.

+++ DemoCutter/demoCut.h
@@ -1985,12 +1985,12 @@ typedef struct {
 
 typedef struct {
 	char* name;
-	int		offset;
+	size_t		offset;
 	int		bits;		// 0 = float
 } netField_t;
 
 // using the stringizing operator to save typing...
-#define	NETF(x) #x,(int)&((entityState_t*)0)->x
+#define	NETF(x) #x,(size_t)&((entityState_t*)0)->x
 
 // using the stringizing operator to save typing...
-#define	PSF(x) #x,(int)&((playerState_t*)0)->x
+#define	PSF(x) #x,(size_t)&((playerState_t*)0)->x
